+++ apps/generator/lib/utils.js
@@ -135,21 +135,17 @@ utils.isAsyncFunction = (fn) => {
 };
 
 /**
- * Register `source-map-support` package.
- * This package provides source map support for stack traces in Node - also for transpiled code from TS.
- *
- * @private
- */
-utils.registerSourceMap = () => {
-  require('source-map-support').install();
-};
-
-/**
  * Register TypeScript transpiler. It enables transpilation of TS filters and hooks on the fly.
  *
  * @private
  */
-utils.registerTypeScript = () => {
+utils.registerTypeScript = (filePath) => {
+  const isTypescriptFile = filePath.endsWith('.ts');
+  
+  if (!isTypescriptFile) {
+    return;
+  }
+  
   const { REGISTER_INSTANCE, register } = require('ts-node');
   // if the ts-node has already been registered before, do not register it again.
   // Check the env. TS_NODE_ENV if ts-node started via ts-node-dev package